fix: frontend knip
